Add GetAlertsByFilter ApiSvcClient method
